test(wallet-core): consolidate reducer tests
