+++ suite-common/wallet-core/src/transactions/__fixtures__/transactionConstants.ts
